apply new test fix to CardGrid tests
