--- src/components/home-v1/CardGrid/Card.test.tsx
@@ -1,58 +1,0 @@
-import * as React from "react";
-import { render, screen, waitFor } from "@testing-library/react";
-import { axe, toHaveNoViolations } from "jest-axe";
-import "@testing-library/jest-dom/extend-expect";
-expect.extend(toHaveNoViolations);
-import renderer from "react-test-renderer";
-// Component
-import Card from "./Card";
-
-const item = {
-  id: "test-id-1",
-  title: "Test",
-  image: {
-    id: "test-id-1",
-    alt: "image-1-alt-text",
-    layout: "responsive",
-    width: 400,
-    height: 200,
-    quality: 90,
-    uri: "https://placeimg.com/400/200/arch",
-    useTransformation: false,
-    transformationLabel: "2_1_960",
-  },
-  url: "https://www.nypl.org/",
-  location: "Stephen A. Schwarzman Building",
-};
-
-describe("Card tests", () => {
-  it("should pass axe accessibility tests", async () => {
-    const { container } = render(<Card item={item} />);
-    await waitFor(async () =>
-      expect(await axe(container)).toHaveNoViolations()
-    );
-  });
-  it("should render a h3 with the title", async () => {
-    render(<Card item={item} />);
-    await waitFor(() =>
-      expect(
-        screen.getByRole("heading", { name: /Test/i, level: 3 })
-      ).toBeInTheDocument()
-    );
-  });
-  // Image should have alt text or have role="none"/"presentatio"
-  // it("should render an image", () => {});
-  it("should add a aria-describedby with additional information", async () => {
-    render(<Card item={item} />);
-    await waitFor(() =>
-      expect(screen.getByRole("heading", { name: /Test/i })).toHaveAttribute(
-        "aria-describedby",
-        "test-id-1-location"
-      )
-    );
-  });
-  it("should render the UI snapshot correctly", async () => {
-    const basicCard = renderer.create(<Card item={item} />).toJSON();
-    await waitFor(() => expect(basicCard).toMatchSnapshot());
-  });
-});
